Update build pipeline and enhance functionality
- Changed the build pipeline to use Rollup
- Updated dependencies
- Automatically add the CDN script if not already present
- Updated types, including handling `window.Frames` being potentially undefined
- Added deprecation warning for future migration to Flow
