+++ .github/workflows/ci.yml
@@ -1,15 +1,26 @@
 name: Test
 
-on: [push]
+on:
+  push:
+    branches:
+      - "**"
 
 jobs:
-    test:
-        runs-on: ubuntu-latest
-        steps:
-            - uses: actions/checkout@v2
-            - uses: actions/setup-node@v2.5.1
-              with:
-                  node-version: 18
-                  registry-url: https://registry.npmjs.org/
-            - run: npm ci --force
-            - run: npm test
+  test:
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout Repository
+        uses: actions/checkout@v2
+
+      - name: Setup Node.js
+        uses: actions/setup-node@v2.5.1
+        with:
+          node-version: 18
+          registry-url: https://registry.npmjs.org/
+
+      - name: Install Dependencies
+        run: npm ci --force
+
+      - name: Run Tests
+        run: npm test
