+++ .gitignore
@@ -5,11 +5,8 @@
 /.pnp
 .pnp.js
 
-# testing
-/coverage
-
 # production
-/build
+/dist
 
 # misc
 .DS_Store
@@ -21,4 +18,4 @@
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
-
+example/
