--- .prettierrc
@@ -1,7 +1,0 @@
-{
-    "semi": true,
-    "singleQuote": true,
-    "tabWidth": 4,
-    "useTabs": false,
-    "printWidth": 100
-}
