+++ package.json
@@ -1,43 +1,46 @@
 {
-    "name": "frames-react",
-    "version": "1.1.2",
-    "peerDependencies": {
-        "react": ">=16.8"
-    },
-    "scripts": {
-        "start": "react-scripts start",
-        "test": "react-scripts test a",
-        "build": "tsc"
-    },
-    "eslintConfig": {
-        "extends": "react-app"
-    },
-    "browserslist": {
-        "production": [
-            ">0.2%",
-            "not dead",
-            "not op_mini all"
-        ],
-        "development": [
-            "last 1 chrome version",
-            "last 1 firefox version",
-            "last 1 safari version"
-        ]
-    },
-    "devDependencies": {
-        "@testing-library/jest-dom": "^6.4.2",
-        "@testing-library/react": "^14.3.0",
-        "@testing-library/user-event": "^14.5.2",
-        "@types/node": "^20.12.6",
-        "@types/react": "^18.2.75",
-        "@types/react-dom": "^18.2.24",
-        "classnames": "^2.5.1",
-        "react": "^18.2.0",
-        "react-dom": "^18.2.0",
-        "react-scripts": "^5.0.1",
-        "typescript": "^5.4.4"
-    },
-    "dependencies": {
-        "classnames": "^2.5.1"
-    }
+  "name": "frames-react",
+  "version": "1.2.0",
+  "main": "dist/cjs/index.js",
+  "scripts": {
+    "build": "rollup -c",
+    "test": "jest",
+    "dev": "npm run build && rm -rf example/src/dist && cp -R dist example/src/dist && cd example && npm install && npm start"
+  },
+  "keywords": [],
+  "license": "MIT",
+  "description": "",
+  "dependencies": {
+    "classnames": "^2.5.1",
+    "react": "^17.0.0 || ^18.0.0",
+    "react-dom": "^17.0.0 || ^18.0.0",
+    "tslib": "^2.8.1"
+  },
+  "devDependencies": {
+    "@rollup/plugin-commonjs": "^28.0.1",
+    "@rollup/plugin-node-resolve": "^15.3.0",
+    "@rollup/plugin-typescript": "^12.1.1",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.0.1",
+    "@types/classnames": "^2.3.4",
+    "@types/jest": "^29.5.14",
+    "@types/react": "^18.3.12",
+    "@types/react-dom": "^18.3.1",
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
+    "rollup": "^2.79.2",
+    "rollup-plugin-peer-deps-external": "^2.2.4",
+    "rollup-plugin-terser": "^7.0.2",
+    "ts-jest": "^29.2.5",
+    "typescript": "^5.6.3"
+  },
+  "peerDependencies": {
+    "react": "^17.0.0 || ^18.0.0",
+    "react-dom": "^17.0.0 || ^18.0.0"
+  },
+  "module": "dist/esm/index.js",
+  "types": "dist/index.d.ts",
+  "files": [
+    "dist"
+  ]
 }
