+++ src/components/CardNumber.tsx
@@ -1,8 +1,8 @@
-import React from 'react';
-import classNames from 'classnames';
-import { CARD_NUMBER_FRAME } from '../config/config';
+import React from "react";
+import classNames from "classnames";
+import { CARD_NUMBER_FRAME } from "../config/config";
 
 export const CardNumber: React.FC<React.HTMLAttributes<HTMLDivElement>> = ({
-    className,
-    ...props
+  className,
+  ...props
 }) => <div className={classNames(CARD_NUMBER_FRAME, className)} {...props} />;
