+++ src/components/ExpiryDate.tsx
@@ -1,8 +1,8 @@
-import React from 'react';
-import classNames from 'classnames';
-import { EXPIRY_DATE_FRAME } from '../config/config';
+import React from "react";
+import classNames from "classnames";
+import { EXPIRY_DATE_FRAME } from "../config/config";
 
 export const ExpiryDate: React.FC<React.HTMLAttributes<HTMLDivElement>> = ({
-    className,
-    ...props
+  className,
+  ...props
 }) => <div className={classNames(EXPIRY_DATE_FRAME, className)} {...props} />;
