--- src/frames/index.tsx
@@ -1,8 +1,0 @@
-export * from './types/types';
-export * from './components/card-number';
-export * from './components/expiry-date';
-export * from './components/cvv';
-export * from './components/card-frame';
-export * from './components/scheme-choice';
-export * from './config/config';
-export * from './Frames';
