+++ src/index.ts
@@ -1,0 +1,7 @@
+export { Frames } from "./components/Frames";
+export { CardFrame } from "./components/CardFrame";
+export { CardNumber } from "./components/CardNumber";
+export { Cvv } from "./components/Cvv";
+export { ExpiryDate } from "./components/ExpiryDate";
+export { SchemeChoice } from "./components/SchemeChoice";
+export * from "./types/types";
